requirements.txt
autopep8==1.5.5
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
idna==2.10
pip-autoremove==0.9.1
pycodestyle==2.6.0
python-decouple==3.4
requests==2.25.1
soupsieve==2.1
toml==0.10.2
urllib3==1.26.3
